refactor: rename src/store to src/state
- change alias in webpack config, tsconfig.json and jest.config.js;
- move src/store/src to src/utils;
- add alias Utils
