+++ jest.config.js
@@ -16,8 +16,9 @@ module.exports = {
     '^Data/(.*)$': '<rootDir>/src/assets/data/$1',
     '^Entities/(.*)$': '<rootDir>/src/entities/$1',
     '^Images/(.*)$': '<rootDir>/src/assets/images/$1',
-    '^Store/(.*)$': '<rootDir>/src/store/$1',
+    '^State/(.*)$': '<rootDir>/src/state/$1',
     '^Styles/(.*)$': '<rootDir>/src/styles/$1',
+    '^Utils/(.*)$': '<rootDir>/src/utils/$1',
   },
   roots: ['<rootDir>/src'],
   setupFilesAfterEnv: ['<rootDir>/src/setupTests.ts'],
