+++ src/components/EditTask/EditTask.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
 import * as React from 'react';
 import { Provider } from 'react-redux';
-import store from 'Store/store';
+import store from 'State/store';
 
 import ITask from 'Entities/task-entities';
 import EditTask from './EditTask';
