+++ src/state/project/reducers.ts
@@ -1,7 +1,7 @@
 import IProject from 'Entities/project-entities';
 import IProjects from 'Entities/projects-entity';
 import * as Types from 'Entities/types';
-import * as MyModels from 'Store/types';
+import * as MyModels from 'State/types';
 import {
   ADD_PROJECT,
   DELETE_PROJECT,
