+++ src/state/task-list/reducers.ts
@@ -1,5 +1,5 @@
 import ITask from 'Entities/task-entities';
-import * as MyModels from 'Store/types';
+import * as MyModels from 'State/types';
 import { DESELECT_TASK, HIDE_EDIT, SELECT_TASK, SHOW_EDIT } from './action-constants';
 import { TaskListState } from './action-types';
