+++ src/state/task/reducers.ts
@@ -1,6 +1,6 @@
 import ITask from 'Entities/task-entities';
 import * as Types from 'Entities/types';
-import * as MyModels from 'Store/types';
+import * as MyModels from 'State/types';
 import {
   ADD,
   ADD_FAILURE,
