+++ webpack.config.prod.js
@@ -24,8 +24,9 @@ module.exports = {
       Data: path.resolve(__dirname, './src/assets/data'),
       Entities: path.resolve(__dirname, './src/entities'),
       Images: path.resolve(__dirname, './src/assets/images'),
-      Store: path.resolve(__dirname, './src/store'),
+      State: path.resolve(__dirname, './src/state'),
       Styles: path.resolve(__dirname, './src/styles'),
+      Utils: path.resolve(__dirname, './src/utils'),
     },
     extensions: ['.ts', '.tsx', '.js', '.jsx', '.json', '.svg'],
   },
@@ -130,7 +131,7 @@ module.exports = {
         test: /\.svg/,
         type: 'asset/inline',
         generator: {
-          dataUrl: content => {
+          dataUrl: (content) => {
             content = content.toString();
             return svgToMiniDataURI(content);
           },
